+++ test/td_i18n/messages_test.exs
@@ -1,6 +1,7 @@
 defmodule TdI18n.MessagesTest do
   use TdI18n.DataCase
 
+  alias TdCache.I18nCache
   alias TdI18n.Messages
   alias TdI18n.Messages.Message
 
@@ -28,7 +29,7 @@ defmodule TdI18n.MessagesTest do
   end
 
   test "create_message/1 with valid data creates a message" do
-    %{id: locale_id} = locale = insert(:locale)
+    %{id: locale_id, lang: lang} = locale = insert(:locale)
 
     %{
       definition: definition,
@@ -44,16 +45,20 @@ defmodule TdI18n.MessagesTest do
              description: ^description,
              message_id: ^message_id
            } = message
+
+    assert ^definition = I18nCache.get_definition(lang, message_id)
+    on_exit(fn -> I18nCache.delete(lang) end)
   end
 
   test "create_message/1 with invalid data returns error changeset" do
-    locale = insert(:locale)
+    %{lang: lang} = locale = insert(:locale)
     params = %{definition: nil, description: nil}
     assert {:error, %Ecto.Changeset{}} = Messages.create_message(locale, params)
+    on_exit(fn -> I18nCache.delete(lang) end)
   end
 
   test "update_message/2 with valid data updates the message" do
-    message = insert(:message)
+    %{message_id: message_id, locale: %{lang: lang}} = message = insert(:message)
 
     update_attrs = %{
       definition: "some updated definition",
@@ -61,8 +66,12 @@ defmodule TdI18n.MessagesTest do
     }
 
     assert {:ok, %Message{} = message} = Messages.update_message(message, update_attrs)
+
     assert message.definition == "some updated definition"
     assert message.description == "some updated description"
+
+    assert message.definition == I18nCache.get_definition(lang, message_id)
+    on_exit(fn -> I18nCache.delete(lang) end)
   end
 
   test "update_message/2 with invalid data returns error changeset" do
@@ -72,8 +81,16 @@ defmodule TdI18n.MessagesTest do
   end
 
   test "delete_message/1 deletes the message" do
-    message = insert(:message)
+    %{message_id: message_id, definition: definition, locale: %{lang: lang}} =
+      message = insert(:message)
+
+    I18nCache.put(lang, %{message_id: message_id, definition: definition})
+
     assert {:ok, %Message{}} = Messages.delete_message(message)
+
     assert_raise Ecto.NoResultsError, fn -> Messages.get_message!(message.id) end
+
+    assert is_nil(I18nCache.get_definition(lang, message_id))
+    on_exit(fn -> I18nCache.delete(lang) end)
   end
 end